cleaned files + added MaskEmbedder + CoordinateEmbedder support in all layers
