+++ stableclimgen/src/models/mgno_transformer/mgno_encoderdecoder_block.py
@@ -1,0 +1,224 @@
+from typing import List,Dict
+import torch.nn as nn
+import torch
+
+from ...modules.icon_grids.grid_layer import GridLayer
+from ...utils.helpers import check_get_missing_key
+from ...modules.neural_operator.no_helpers import get_no_layer,get_embedder_from_dict,get_embedder
+from ...modules.neural_operator.no_blocks import PreActivation_NOBlock, NOBlock
+from ...modules.neural_operator import mg_layers as mg
+
+class MGNO_EncoderDecoder_Block(nn.Module):
+  
+    def __init__(self,
+                 input_levels: List[int],
+                 input_dims: List[int],
+                 global_levels_decode: List[int],
+                 global_levels_no: List[int],
+                 model_dims_out: List[int],
+                 grid_layers: List[GridLayer],
+                 layer_settings: dict,
+                 rotate_coordinate_system: bool = True,
+                 rule = 'fc', # ">" "<"
+                 mg_reduction = 'linear',
+                 mg_reduction_embed_confs: Dict = None,
+                 mg_reduction_embed_names: List = None,
+                 mg_reduction_embed_names_mlp: List = None,
+                 mg_reduction_embed_mode: str = 'sum',
+                 mg_att_dim = 128,
+                 mg_n_head_channels = 16,
+                 p_dropout=0,
+                 mask_as_embedding = False,
+                 level_diff_zero_linear = False
+                ) -> None: 
+      
+        super().__init__()
+
+        self.mask_as_embedding = mask_as_embedding
+        self.output_levels = global_levels_decode
+        self.model_dims_out = model_dims_out
+
+        self.layers = nn.ModuleList()
+        self.reduction_layers = nn.ModuleList()
+        self.module_indices = []
+        self.output_indices = []
+        self.layer_indices = []
+        self.add_coordinate_embedding_dict = {}
+        layer_index = 0
+        for output_idx, output_level in enumerate(global_levels_decode):
+
+            input_indices = []
+            layer_indices = []
+            mg_input_dims = []
+            mg_input_levels = []
+            for input_idx, input_level in enumerate(input_levels):
+
+                level_diff = output_level - input_level
+
+                if rule == "<" and level_diff>0:
+                   # self.layers.append(nn.Identity())
+                    continue
+
+                elif ">" in rule and level_diff<0:
+                    continue
+
+                if rule == ">max" and input_level!=max(input_levels) and level_diff!=0:
+                    continue
+                
+                input_indices.append(input_idx)
+
+                if list(layer_settings.keys())[0].isnumeric():
+                    layer_setting = layer_settings[str(level_diff)]
+                else:
+                    layer_setting = layer_settings
+                
+                model_dim_in = input_dims[input_idx]
+                model_dim_out = model_dims_out[output_idx]
+                mg_input_dims.append(model_dim_out)
+                mg_input_levels.append(input_level)
+
+                global_level_no = global_levels_no[output_idx]
+
+                layer_type = check_get_missing_key(layer_setting, "type")
+
+                if level_diff_zero_linear and level_diff==0:
+                    layer = nn.Linear(model_dim_in, model_dim_out) if model_dim_in!=model_dim_out else nn.Identity()
+
+                elif 'NO' in layer_type:
+
+                    no_layer_settings = check_get_missing_key(layer_setting, "no_layer_settings")
+                    no_layer_type = check_get_missing_key(layer_setting, "no_layer_type")
+
+                    no_layer = get_no_layer(no_layer_type,
+                                            grid_layers[str(input_level)],
+                                            grid_layers[str(global_level_no)],
+                                            grid_layers[str(output_level)],
+                                            precompute_encode=True,
+                                            precompute_decode=True,
+                                            rotate_coordinate_system=rotate_coordinate_system,
+                                            layer_settings=no_layer_settings,
+                                            normalize_to_mask= (mask_as_embedding==False))
+                
+                    if "mlp" in layer_type:
+                        embedder = get_embedder_from_dict(layer_setting)
+                        if 'pre' in layer_setting["type"]:
+                            layer = PreActivation_NOBlock(
+                                        model_dim_in=model_dim_in,
+                                        model_dim_out=model_dim_out,
+                                        no_layer=no_layer,
+                                        embedder=embedder,
+                                        cross_no = 'cross_no' in layer_setting["type"],
+                                        with_gamma = 'gamma' in layer_setting["type"],
+                                        mask_as_embedding=mask_as_embedding
+                                        )
+                        else:
+                            layer = NOBlock(
+                                        model_dim_in=model_dim_in,
+                                        model_dim_out=model_dim_out,
+                                        no_layer=no_layer,
+                                        embedder=embedder,
+                                        cross_no = 'cross_no' in layer_setting["type"],
+                                        with_gamma = 'gamma' in layer_setting["type"],
+                                        mask_as_embedding=mask_as_embedding
+                                        )
+                elif 'linear' in layer_type:
+                    layer = nn.Linear(model_dim_in, model_dim_out) if model_dim_in!=model_dim_out else nn.Identity()
+                
+                self.layers.append(layer)
+                layer_indices.append(layer_index)
+                layer_index += 1
+
+            if len(mg_input_dims)>1:
+                if mg_reduction == 'linear':
+                    reduction_layer = mg.LinearReductionLayer(mg_input_dims, model_dim_out)
+                
+                elif mg_reduction == 'sum':
+                    reduction_layer = mg.SumReductionLayer()
+
+                elif mg_reduction == 'MGAttention' or 'MGDiffAttention'or 'MGDiffMAttention':
+                    if mg_reduction_embed_names is not None:
+                        embedder = get_embedder(embed_names=mg_reduction_embed_names,
+                                                embed_confs=mg_reduction_embed_confs,
+                                                embed_mode=mg_reduction_embed_mode)
+                        
+                    else:
+                        embedder = None
+                    if mg_reduction_embed_names_mlp is not None:
+                        embedder_mlp = get_embedder(embed_names=mg_reduction_embed_names_mlp,
+                                                embed_confs=mg_reduction_embed_confs,
+                                                embed_mode=mg_reduction_embed_mode)
+                    else:
+                        embedder_mlp = None
+                    
+                    if mg_reduction == 'MGAttention':
+                        reduction_layer = mg.MGAttentionReductionLayer(mg_input_levels,
+                                                                    mg_input_dims, 
+                                                                    model_dim_out,
+                                                                    att_dim=mg_att_dim,
+                                                                    n_head_channels=mg_n_head_channels,
+                                                                    embedder_grid=embedder,
+                                                                    embedder_mlp=embedder_mlp,
+                                                                    p_dropout=p_dropout)
+                    elif mg_reduction == 'MGDiffAttention':
+                        reduction_layer = mg.MGDiffAttentionReductionLayer(torch.tensor(mg_input_levels)-output_level,
+                                                                    mg_input_dims, 
+                                                                    model_dim_out,
+                                                                    att_dim=mg_att_dim,
+                                                                    n_head_channels=mg_n_head_channels,
+                                                                    embedder_grid=embedder,
+                                                                    embedder_mlp=embedder_mlp,
+                                                                    p_dropout=p_dropout)
+                    elif mg_reduction == 'MGDiffMAttention':
+                        reduction_layer = mg.MGDiffMAttentionReductionLayer(torch.tensor(mg_input_levels)-output_level,
+                                                                    mg_input_dims, 
+                                                                    model_dim_out,
+                                                                    att_dim=mg_att_dim,
+                                                                    n_head_channels=mg_n_head_channels,
+                                                                    embedder_grid=embedder,
+                                                                    embedder_mlp=embedder_mlp,
+                                                                    p_dropout=p_dropout)
+            else:
+                reduction_layer = mg.IdentityReductionLayer()
+            
+            self.reduction_layers.append(reduction_layer)
+
+            self.layer_indices.append(layer_indices)
+            self.output_indices.append(input_indices)
+    
+    
+
+    def forward(self, x_levels, coords_in=None, coords_out=None, indices_sample=None, mask_levels=None, emb=None):
+
+        x_levels_out = []
+        mask_levels_out = []
+
+        for output_index, input_indices in enumerate(self.output_indices):
+
+            outputs_ = []
+            masks_ = []
+
+            for layer_index, input_index in enumerate(input_indices):
+                x = x_levels[input_index]
+                mask = mask_levels[input_index]
+                
+
+                layer = self.layers[self.layer_indices[output_index][layer_index]]
+                
+                if isinstance(layer, nn.Identity) or isinstance(layer, nn.Linear):
+                    x_out = layer(x)
+                    mask_out = mask
+                else:
+                    x_out, mask_out = layer(x, coords_encode=coords_in, coords_decode=coords_out, indices_sample=indices_sample, mask=mask, emb=emb)
+
+                if mask_out is not None:
+                    mask_out = mask_out.view(x_out.shape[:3])
+
+                masks_.append(mask_out)
+                outputs_.append(x_out)
+
+            x_out, mask_out = self.reduction_layers[output_index](outputs_, mask_levels=masks_, emb=emb)
+
+            x_levels_out.append(x_out)
+            mask_levels_out.append(mask_out)
+
+        return x_levels_out, mask_levels_out
