+++ stableclimgen/src/modules/embedding/embedder.py
@@ -59,27 +59,6 @@ class CoordinateEmbedder(BaseEmbedder):
             torch.nn.Linear(self.embed_dim, self.embed_dim),
         )
 
-class MaskEmbedder(BaseEmbedder):
-    """
-    A neural network module to embed longitude and latitude coordinates.
-
-    :param embed_dim: Dimensionality of the embedding output.
-    :param in_channels: Number of input coordinate features (default is 2).
-    """
-
-    def __init__(self, name: str, in_channels: int, embed_dim: int) -> None:
-        super().__init__(name, in_channels, embed_dim)
-
-        # keep batch, spatial, variable and channel dimensions
-        self.keep_dims = ["b", "t", "s", "v", "c"]
-
-        # Mesh embedder consisting of a RandomFourierLayer followed by linear and GELU activation layers
-        self.embedding_fn = torch.nn.Sequential(
-            torch.nn.Linear(self.in_channels, self.embed_dim),
-            torch.nn.GELU(),
-            torch.nn.Linear(self.embed_dim, self.embed_dim),
-        )
-
 class VariableEmbedder(BaseEmbedder):
 
     def __init__(self, name: str, in_channels: int, embed_dim: int, init_value:float = None) -> None:
@@ -94,12 +73,12 @@ class VariableEmbedder(BaseEmbedder):
 
 class MaskEmbedder(BaseEmbedder):
 
-    def __init__(self, name: str,  embed_dim: int, in_channels: int=2, init_value:float = None) -> None:
+    def __init__(self, name: str,  in_channels: int, embed_dim: int, init_value:float = None) -> None:
         super().__init__(name, in_channels, embed_dim)
 
         self.keep_dims = ["b", "s", "v", "c"]
 
-        self.embedding_fn = nn.Embedding(self.in_channels, self.embed_dim)
+        self.embedding_fn = nn.Embedding(2, self.embed_dim)
 
         if init_value is not None:
             self.embedding_fn.weight.data.fill_(init_value)
